pyproject.toml
[project]
name = "mercury-virtual-solar-wind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "astropy>=7.2.0",
    "cdflib>=1.3.8",
    "check-shapes>=1.1.1",
    "gpflow>=2.5.2",
    "hermpy",
    "numpy>=2.4.1",
    "polars>=1.38.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "setuptools<82",
    "sunpy>=7.1.0",
    "sunpy-soar>=1.12.0",
    "tf-keras>=2.20.1",
]

[tool.uv.sources]
hermpy = { git = "https://github.com/daraghhollman/hermpy-1.0.0.git" }

[build-system]
requires = ["setuptools"]

[tool.setuptools.packages.find]
where = ["src"]
